requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx
python-dotenv
requests
